first stab at fitting hierarchical B-spline models
goal is to fit hierachical B-spline model with coeffs for plants drawn centred on coeffs for groups.

have R and stan which compiles, and think runs correctly, but TOO SLOW to be useful! (~15 mins per metric with only 10 knots). Add to git as documents idea, (based on DOI:10.1089/dia.2010.0199) but next step is to try and use brms directly, or figure out how to optimise.